修复日志环形队列的 ABA 问题
head/tail 原为回绕索引，生产者 CAS 在索引转一整圈后仍可能用过期的
满检查成功，造成幽灵空环、ready 标志脏读撕裂日志行甚至消费者卡死。
改为单调递增 64 位计数器取模定位槽位，CAS 值全局唯一，不再复现。
顺带修复：崩溃路径 logger_flush 与轮转中的 FILE* 竞争改为写 .crash
伴生文件；ring_pop 改 trylock 避免信号处理自死锁；关停排空改有界
重试；消费者区分真空闲与队头未就绪，避免误睡导致溢出丢日志。

+++ common/logger.h
@@ -5,6 +5,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <stdarg.h>
+#include <stdint.h>
 #include <time.h>
 #include <pthread.h>
 #include <sys/time.h>
@@ -38,8 +39,11 @@ typedef struct {
 typedef struct {
     log_entry_t *ring;
     int *ready;
-    int head;
-    int tail;
+    /* Monotonic 64-bit counters (never wrap): slot = counter & mask. CAS on a
+     * monotonic counter can't succeed on a stale value (no ABA), unlike wrapped
+     * indices that reappear after a full lap. */
+    uint64_t head;
+    uint64_t tail;
     int size;
     int mask;
     pthread_mutex_t pop_mtx;
